Force-load theme at the right moment
It seems that the group manager is implicitly triggering the loading
routine for the theme namespace, but that one is not available yet at
this stage.

The fix is to force load the theme app so that the namespace is
available early enough when needed.

This fixes an issue where the theme was not appearing for guest users.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>